Created Coding- Sum, Counts, and Average
Created Python code to determine if the user enters 'done' to exit the loop, converts the user input string to a float number or gives an error, counts how many numbers were entered, and sums the numbers to get the total. Finally, the code prints the final total and number of counts and calculates and prints the average by dividing the total by counts.